Update the Unit test
