+++ Example/Tests/HTTPRetrySpec.swift
@@ -15,19 +15,19 @@ import Combine
 class HTTPRetrySpec: QuickSpec {
     // swiftlint:disable function_body_length
     override class func spec() {
-        var sender: URLRequestSenderMock!
-        var publisher: URLSession.HTTPRetry<URLRequestSenderMock>!
+        var sender: MockablePublisher!
+        var publisher: URLSession.HTTPRetry<MockablePublisher>!
         var retrier: MockRetrier!
         context("request is failing") {
             beforeEach {
                 retrier = MockRetrier()
-                sender = URLRequestSenderMock(result: .failure(.expectedError))
-                publisher = .init(sender: sender, retrier: retrier)
+                sender = MockablePublisher(.failure(.error(TestError.expectedError)))
+                publisher = .init(upstream: sender, retrier: retrier)
             }
             it("should drop the request") {
                 retrier.decision = .drop
                 
-                let result = try sendRequest(using: publisher)
+                let result = try waitForResponse(to: publisher)
                 
                 expect(retrier.called).to(beTrue())
                 expectToBeExpectedError(for: result)
@@ -35,7 +35,7 @@ class HTTPRetrySpec: QuickSpec {
             it("should drop the request with reason") {
                 retrier.decision = .dropWithReason(reason: "expected error")
                 
-                let result = try sendRequest(using: publisher)
+                let result = try waitForResponse(to: publisher)
                 
                 expect(retrier.called).to(beTrue())
                 expectToBeToRetryError(for: result)
@@ -43,32 +43,22 @@ class HTTPRetrySpec: QuickSpec {
             it("should retry the request") {
                 retrier.decision = .retry
                 
-                let result = try sendRequest(using: publisher)
+                let result = try waitForResponse(to: publisher)
                 
                 expect(retrier.called).to(beTrue())
-                expectToBeWhileRetryError(for: result)
-            }
-            it("should retry with new request") {
-                let newRequest = URLRequest(url: URL(string: "http://www.adapt.com")!)
-                retrier.decision = .retryWithNewRequest(newRequest)
-                
-                let result = try sendRequest(using: publisher)
-                
-                expect(sender.sentRequest).to(equal(newRequest))
-                expect(retrier.called).to(beTrue())
-                expectToBeWhileRetryError(for: result)
+                expectToBeExpectedError(for: result)
             }
         }
         context("request is succeed") {
             beforeEach {
                 retrier = MockRetrier()
-                sender = URLRequestSenderMock(result: .success((Data(), HTTPURLResponse())))
-                publisher = .init(sender: sender, retrier: retrier)
+                sender = MockablePublisher(.success((data: Data(), response: HTTPURLResponse())))
+                publisher = .init(upstream: sender, retrier: retrier)
             }
             it("should success and never drop") {
                 retrier.decision = .drop
                 
-                let result = try sendRequest(using: publisher)
+                let result = try waitForResponse(to: publisher)
                 
                 expect(retrier.called).to(beFalse())
                 expectToBeDefaultSuccess(for: result)
@@ -76,7 +66,7 @@ class HTTPRetrySpec: QuickSpec {
             it("should success and never drop with reason") {
                 retrier.decision = .dropWithReason(reason: "unexpected error")
                 
-                let result = try sendRequest(using: publisher)
+                let result = try waitForResponse(to: publisher)
                 
                 expect(retrier.called).to(beFalse())
                 expectToBeDefaultSuccess(for: result)
@@ -84,16 +74,7 @@ class HTTPRetrySpec: QuickSpec {
             it("should success and never retry") {
                 retrier.decision = .retry
                 
-                let result = try sendRequest(using: publisher)
-                
-                expect(retrier.called).to(beFalse())
-                expectToBeDefaultSuccess(for: result)
-            }
-            it("should success and never retry with new reqeust") {
-                let newRequest = URLRequest(url: URL(string: "http://www.adapt.com")!)
-                retrier.decision = .retryWithNewRequest(newRequest)
-                
-                let result = try sendRequest(using: publisher)
+                let result = try waitForResponse(to: publisher)
                 
                 expect(retrier.called).to(beFalse())
                 expectToBeDefaultSuccess(for: result)
@@ -101,32 +82,6 @@ class HTTPRetrySpec: QuickSpec {
         }
     }
     // swiftlint:enable function_body_length
-}
-
-// MARK: Test
-
-private func sendRequest(using publisher: URLSession.HTTPRetry<URLRequestSenderMock>) throws -> Result<(data: Data, response: URLResponse), HTTPURLError> {
-    var result: Result<(data: Data, response: URLResponse), HTTPURLError>?
-    var cancellable: AnyCancellable?
-    waitUntil { done in
-        cancellable = publisher.sink { completion in
-            switch completion {
-            case .finished:
-                break
-            case .failure(let error):
-                result = .failure(error)
-            }
-            done()
-        } receiveValue: { value in
-            result = .success(value)
-        }
-    }
-    cancellable?.cancel()
-    guard let result else {
-        fail("result should not be nil at this point")
-        throw TestError.unexpectedError
-    }
-    return result
 }
 
 // MARK: Expectation
@@ -145,7 +100,7 @@ private func expectToBeWhileRetryError(for result: Result<(data: Data, response:
     case .success:
         fail("result should fail")
     case .failure(let error):
-        guard case .failWhileRetry(let underlyingError, _, _) = error,
+        guard case .failWhileRetry(let underlyingError, _) = error,
               let httpUrlError = underlyingError as? HTTPURLError,
               case .error(let subUnderlyingError) = httpUrlError,
                 let testError = subUnderlyingError as? TestError else {
@@ -161,7 +116,7 @@ private func expectToBeToRetryError(for result: Result<(data: Data, response: UR
     case .success:
         fail("result should fail")
     case .failure(let error):
-        guard case .failToRetry(let reason, _, _) = error else {
+        guard case .failToRetry(let reason, _) = error else {
             fail("result should produce TestError but produce \(String(describing: error))")
             return
         }
@@ -185,13 +140,15 @@ private func expectToBeExpectedError(for result: Result<(data: Data, response: U
 
 private class MockRetrier: HTTPDataTaskRetrier {
     
-    var decision: HTTPDataTaskRetryDecision = .drop
-    var called: Bool = false
     
-    func httpDataTaskShouldRetry(for error: HTTPURLError, request: URLRequest) async throws -> HTTPDataTaskRetryDecision {
+    var decision: HTTPDataTaskRetryDecision = .drop
+    var called: Bool { calledCount > 1 }
+    var calledCount: Int = 0
+    
+    func httpDataTaskShouldRetry(for error: HTTPURLError) async throws -> HTTPDataTaskRetryDecision {
         // make sure only retry once so the test will not retry forever
         guard !called else { return .drop }
-        called = true
+        calledCount  += 1
         return decision
     }
 }
